Moved and renamed utils to DateTime
